ajout code gpu

+++ gadgets/mri_core/GenericReconCartesianGrappaGadget.cpp
@@ -424,9 +424,9 @@ namespace Gadgetron {
             // only allow this for loop openmp if num>1 and 2D recon
 #pragma omp parallel for default(none) private(ii) shared(src, dst, recon_obj, e, num, ref_N, ref_S, ref_RO, ref_E1, ref_E2, RO, E1, E2, dstCHA, srcCHA, convKRO, convKE1, convKE2, kRO, kNE1, kNE2) if(num>1)
             for (ii = 0; ii < num; ii++) {
-                size_t slc = ii / (ref_N * ref_S);
-                size_t s = (ii - slc * ref_N * ref_S) / (ref_N);
-                size_t n = ii - slc * ref_N * ref_S - s * ref_N;
+                long long slc = ii / (ref_N * ref_S);
+                long long s = (ii - slc * ref_N * ref_S) / (ref_N);
+                long long n = ii - slc * ref_N * ref_S - s * ref_N;
 
                 std::stringstream os;
                 os << "n" << n << "_s" << s << "_slc" << slc << "_encoding_" << e;
